Downgrade autofac dependency to work with pre release v0.0.55-pre25
