+++ TougePlugin/lua/hud.lua
@@ -1,7 +1,5 @@
 local baseUrl = "http://" .. ac.getServerIP() .. ":" .. ac.getServerPortHTTP() .. "/static/TougePlugin/"
 
-local scaling = require('scaling')
-
 local windowWidth = sim.windowWidth
 local windowHeight = sim.windowHeight
 
@@ -45,6 +43,25 @@ local hasIncomingNotification = false
 local notificationActivatedAt = nil
 
 local sim = ac.getSim()
+
+-- Scaling
+local baseRes = vec2(2560, 1440) -- Reference resolution
+local currentRes = vec2(sim.windowWidth, sim.windowHeight)
+local scaleFactor = math.min(currentRes.x / baseRes.x, currentRes.y / baseRes.y, 1)
+
+local scaling = {}
+
+function scaling.vec2(x, y)
+  return vec2(x, y) * scaleFactor
+end
+
+function scaling.size(size)
+  return size * scaleFactor
+end
+
+function scaling.get()
+  return scaleFactor
+end
 
 -- Load fonts
 local fontsURL = baseUrl .. "fonts.zip"
